Byte-compilation trick
This allows applications of `pure' functions as well as literal
objects in `values' to propagate and compile to a literal circular
list.  Unfortunately the code is made less efficient when any argument
in `values' is non-constant, but I consider this the less common case.
The original implementation also contains a bug regarding `quote',
which is resolved here.

* for-sequence.el (for--make-circular, for-in-repeat):