Remove `for-binder'
This variable doesn't work properly because the forms cannot be
guaranteed to be expanded during the dynamic extent of the desired
binding, which is apparent in two cases:

1. Interpreted functions when eager expansion is disabled;
2. Interpreted Lisp files where file/directory-local variables aren't
read in.

As a result, I think it's best to remove this variable entirely.

* for-iteration.el
(for-binder): Remove variable and always use `pcase-let'.
(for-iteration-clause, for-special-clause): Follow `pcase-let'.
* for.texi: Document the change.
* for-tests.el: Remove test cases.